fix(devtools-view): Fix selectable menu section role (#23177)
[Previous PR](https://github.com/microsoft/FluidFramework/pull/23093)
attempted to fix this by setting the "button" role via style properties,
but apparently this doesn't actually work. Confirmed that the role did
not appear in the browser devtools view previously but does with the new
pattern.

Also updates styling to highlight selectable menu headers the same way
we highlight sub-heading selections.