feat(build-infrastructure): Add selection and filtering APIs (#22866)
This change adds selection and filtering logic to the
build-infrastructure package. Much of this code currently exists in
build-cli, but the implementation has changed slightly to accommodate
the new interfaces. There are also new git-related APIs, which are used
in the "changed since" selection/filter logic.

None of the code is shared with existing implementation in build-cli.
Once build-infrastructure is ready, the build-cli functionality will be
replaced with the build-infra stuff.

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>