fix(gc): fullGC flag for Autorecovery shouldn't be cleared before GC runs (#23179)
When reviewing the telemetry for autorecovery, I found that a large
number of documents see Autorecovery but not fullGC. This was unexpected
since Autorecovery includes setting a flag that causes the next GC run
to use fullGC.

Upon closer inspection of a few sample sessions, I found that the flag
was getting cleared before the next GC run. Likely due to unexpected
case where summary triggers the load of the Tombstoned object, so the GC
op and the summary Ack are racing. If the GC op wins, then the summary
Ack clears the flag.

This updates the logic to only clear the flag if GC actually ran with
fullGC.