feat(eslint-config-fluid): Add strict-biome config for use in conjunction with Biome linter (#22960)
Adds a strict config to the shared eslint package that can be used in
projects to use Biome for linting in conjunction with eslint. This
config is the same as the strict config, but disables rules that have a
Biome equivalent.

To make it easier to see the differences, I also updated the strict
config - this __will not__ be merged.