Skip flaky SharedString e2e test for ODSP and r11s (#23187)
This test is observed to be somewhat flaky in the e2e pipeline. There
isn't much to be done for this test as it's testing simple behavior, and
the flakiness is most likely due to server performance at the time of
the pipeline run. Since the test is simply verifying end-to-end flow
(loader -> runtime -> DDS) for a specific feature, we can skip it for
the specifically flaky drivers.


[AB#12905](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/12905)

[AB#13656](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/13656)

Data point side note when running the test 1000 times for specific
drivers:
- ODSP 0/1000 failed
- FRS 2/1000 failed