feat(odsp-driver): Add logging property if doc is in dogfood or MSIT server environments (#23186)
Sometimes we are investigating a live site incident with our 1P
customers, and it would be helpful to know which service environment
(SharePoint "Farm") the document belongs to, since different settings or
code can be deployed to these independently. We can get a reasonable
proxy of this by inspecting the domain of the URL against some known
patterns.

We are computing/logging this for trees-latest and create-new events -
should typically be once per session.