refactor: Use assert.equal() to see values when test fails (#23172)
## Description

The `Includes ack'd ids in summary` test sometimes fails against
routerlicious but the error output only says `false !== true` so we
can't tell the actual value of the number that is being compared to 1.
This PR updates a few asserts in that file so they'll give us the actual
value instead of just `expected true, got false` in the output when the
test fails.