fix(examples): Fix tests in external-data example so they don't fail in Node20 (#23189)
## Description

Fixes the tests of the external-data example so they succeed in Node20.

Also simplifies that package a bit by removing an unnecessary split of
the webpack config and removing the unused `stream-http` dependency.

I also tried to simplify further by replacing node-fetch with Node's
native fetch but that led to different test failures. Not urgent so I
left it alone.