build(client): Fix configs preventing incremental build (#22897)
Fixes several problems causing builds to not be fully incremental.

1. data-object-base had no tests but included test-related tasks. Since
no test output was generated, the task was never incremental. Those
superfluous tasks have been removed. They can be added back if/when the
package is no longer experimental.
2. Added new declarative tasks for `flub check buildversion` and
`markdown-magic`.
3. The ai-collab project has `"allowJs": true` in its tsconfig. That
setting causes problems with incremental builds (see #23197 for more
details). It's removed in this change but if it is truly we can add it back.

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>