Add nav parameter to the sharing link in ODSP driver resolved url (#22849)
Fixes:
[AB#20323](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/20323)

## Description

When we create a file, we receive the sharing link information from the
service in the response. We then use that to set the sharingLink prop on
the resolved url which is then supposed to be used by the hosts. We
extract the sharing link info in this function:
/**
* If user requested creation of a sharing link along with the creation
of the file by providing
* createLinkScope in the request parameters then extract and save the
sharing link information from
 * the response if it is available.
* In case there was an error in creation of the sharing link, error is
provided back in the response,
 * and does not impact the creation of file in ODSP.
* @param requestedSharingLinkKind - Kind of sharing link requested to be
created along with the creation of file.
* @param response - Response object received from the /snapshot api call
* @returns Sharing link information received in the response from a
successful creation of a file.
 */
function extractShareLinkData(
    response: ICreateFileResponse,
    enableSingleRequestForShareLinkWithCreate?: boolean,
): ShareLinkInfoType | undefined {
    let shareLinkInfo: ShareLinkInfoType | undefined;
    if (enableSingleRequestForShareLinkWithCreate) {
// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
        const { sharing } = response;
        if (!sharing) {
            return;
        }
/* eslint-disable @typescript-eslint/no-unsafe-assignment,
@typescript-eslint/no-unsafe-member-access */
        shareLinkInfo = {
            createLink: {
                link: sharing.sharingLink
                    ? {
                            scope: sharing.sharingLink.scope,
                            role: sharing.sharingLink.type,
                            webUrl: sharing.sharingLink.webUrl,
                            ...sharing.sharingLink,
                        }
                    : undefined,
                error: sharing.error,
                shareId: sharing.shareId,
            },
        };
/* eslint-enable @typescript-eslint/no-unsafe-assignment,
@typescript-eslint/no-unsafe-member-access */
    }
    return shareLinkInfo;
}

However, the hosts also needs the nav parameter as a query parameter on
this link(weburl), we use appendLocatorParams, to append that query
parameter, however that is in the URl resolver class. We can separate
that code into a separate utility api which is called by both the
appendLocatorParams in the resolver and also by the createFile.ts in the
odsp driver in order to append that parameter.