Enabled graceful Redis shutdown in r11 (#23212)
## Description

This PR adds supports for graceful shutdown of Redis client connections.
Prior to this fix, all pod terminations would result in a lot of Redis
client errors. SHutting down connections would prevent this.

## Breaking Changes

Adds a new resource to the `runnerFactory` of `Alfred`, `Nexus` and
`Riddler`. This change could be breaking for FRS and will be fixed by
consuming the latest OSS package into the repo. Added changesets for
these.