Enabled redis graceful shutdown in Gitrest and Historian (#23225)
## Description

This is a follow up PR to
https://github.com/microsoft/FluidFramework/pull/23212. It adds supports
for graceful shutdown of Redis client connections. Prior to this fix,
all pod terminations would result in a lot of Redis client errors.
Shutting down connections would prevent this.

## Breaking Changes

Adds a new resource to the `runnerFactory` of `Historian`. This change
could be breaking for FRS and will be fixed by consuming the latest OSS
package into the repo. It also modifies the `IFileSystemManagerFactory`
interface to support a new method - `.dispose()`. This method will be
used to dispose any connections/resources.