feat: Make NodeFromSchema work in more cases (#23089)
## Description

NodeFromSchema now prefers the class based form if given a schema which
implements a class based and non class based API.

This shouldn't impact our public APIs (unless someone adds a static
create function to their class based schema, in which case this is a
fix), but internally we use TreeNodeSchemaBoth in some places, and if
you subclass a type implementing TreeNodeSchemaBoth, before this change
NodeFromSchema would give the less specialized type from the create
function, ignoring the subclassing. This was confusing and has been
fixed.