refactor(test-tools): Make initial port configurable and update deps (#23202)
## Description

This updates the initial port that test-tools assigns to packages from
8081 to 9000 and makes it configurable.

### Motivation

The build of the LTS branch is failing, and as far as I can tell it is
due to a combination of two things:
- _Something_ (can't figure out what) in the build agent image is
listening on port 8084
- In LTS in particular (compared to main, for example), the package that
ends up getting assigned port 8084 by the `assign-test-ports` executable
from `@fluidframework/test-tools` has jest tests so it tries to start
jest in that port and fails.

I imagine this started failing at some point when we made changes to the
build infra, but since we're not continuously building the LTS branch,
we didn't notice at the time. I even confirmed that running the build
pipeline for the last commit in LTS for which it succeeded (May 2024),
fails today.

The dependency update is purely opportunistic.