Refactor SharedTree transactions to use a branch to track their state (#23241)
## Description

This enables various cleanup, simplification, and encapsulation above
and below the checkout layer of our architecture.

* Logic related to managing the transaction branch is implemented in
`SquashingTransactionStack`, a new `Transactor` implementation (which
extends the current implementation, `TransactionStack`). The complexity
of switching between the original branch and the transaction branch is
handled by the stack so that the checkout does not need to worry about
it.
* `SharedTreeBranchEvents` has been simplified to be "append", "remove"
or "rebase". The "replace" variant, which previously covered both
rebases and transaction commits, is gone. Instead, transaction commits
manifest as an "append" of a squashed commit. This cleans up many of the
event listeners all over the place that had to carefully handle
"replace".
* `TestSharedTreeCore` has been updated to use the new
`SquashingTransactionStack` so that it mirrors the behavior of an actual
`SharedTree` (this allows commit enrichment and resubmission to continue
being unit tested).