[bump] Bump test tools to next major version (#23265)
## Description

Bumps the (independent) test-tools package to the next major version.
This is because we want to release 2.0 (with the changes from
https://github.com/microsoft/FluidFramework/pull/23202) in order to fix
the build of the LTS branch, and the first step towards that is bumping
the package version in main to then cut the release branch from the
commit before the bump is merged.

Note: after this release, I'm hoping to make test-tools a private
package inside the client release group (in
https://github.com/microsoft/FluidFramework/pull/23211), so 2.x would be
the last version we release for it. This bump would be just to satisfy
the process to release it today.