fix(gitrest): ensure there are size limit checks before compression (#23261)
## Description

Currently, when Gitrest checks file size within Isomorphic-git, the file
is already compressed. That means when Gitrest can write a file above
specified limits if it compresses below the limit. This makes limiting
file sizes for the purpose of service memory use inconsistent and
unreliable.

This PR adds an extra config, `git:maxBlobSizeBytes`. This new config is
distinct from the existing `git:filesystem:maxFileSizeBytes` and
`git:ephemeralFilesystem:maxFileSizeBytes` because this pre-compression
limit is about _service memory_, not storage size. The filesystem limits
will still apply to their respective filesystems.

For example, an 300kb Ephemeral summary that does not compress is less
than the maxBlobSize but greater than the maxFileSize for Redis, so it
will still throw an EFBig error. However, a 300kb Ephemeral summary that
compresses to 100Kb will not throw, and this is OK because Redis will
only know about the 100Kb, which is in spec.

## Breaking Changes

The internally used RepoManagerFactoryBase and RepoManagerBase
constructor params has a partial change to an `options` object. This
will help avoid the param list getting overly long and confusing.
However, the publicly used IsomorphicGitManagerFactory simply had an
additional optional param added, so public APIs should be unchanged.