Provide standalone apis for loader class (#22902)
## Description


[AB#26406](https://dev.azure.com/fluidframework/internal/_workitems/edit/26406)


Provide standalone APIs to create and load containers instead of using
the Loader object to do so. Earlier hosts were
supposed to create the Loader object first and then call methods on it
to create and load containers. Now they can just
utilize these APIs and get rid of the Loader object.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>
Co-authored-by: Tony Murphy <anthony.murphy@microsoft.com>
Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>
Co-authored-by: Matt Rakow <ChumpChief@users.noreply.github.com>
Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>