Reduce use of Proxyquire (#23248)
## Description

Proxyquire hacks the behavior require to impact module resolution. This
isn't very portable, and causes issues with ESM.

This change removes the need for such hacks by refactoring the script
who's imports were being overridden into a script and a separate run
function which takes the global it needs as a parameter. The test is
then rewritten against the runt function instead of the script.

This approach, if approved, can be applied to the other scripts in the
package to remove the need for proxyquire, fixing esm support and
resolving
https://dev.azure.com/fluidframework/internal/_workitems/edit/7470