Add IContainerRuntimeOptionInternal for soon-to-be-deprecated options (#23272)
We are working towards removing support for certain
`IContainerRuntimeOptions`. `flushMode.Immediate` and
`enableGroupedBatching` in particular.

We can remove them from the public (meaning legacy-alpha) API now, but
we need to be able to include them internally for compatibility
configuration needed for apps needing to support collaboration between
1.x and 2.x clients.

So this PR adds a new internal interface
`IContainerRuntimeOptionInternal` which has these two options, and uses
it for compat config. We can now deprecate/remove the options from the
main interface.