refactor(build-tools): Deprecate some outdated commands and APIs (#23280)
There are a number of build-tools APIs that should no longer be used for
new code. We also have some commands that are no longer used and can be
removed in the future. To help clarify which APIs and commands are
definitely going away, I have made the following changes:

1. Removed the Timer class export from the build-tools package.
2. Removed use of Timer in the `check:layers` command, including the
logtime flag. The flag was hidden and unused.
3. Updated Logger function signatures and implementations to be
compatible with oclif logging functions, so that the base class can
correctly call the logger used by the concrete instance.
4. Deprecated MonoRepoKind type and removed some uses in test files
(prod code remains unchanged).
5. Deprecated FluidRepo, GitRepo, MonoRepo, readPackageJsonAndIndent,
updatePackageJsonFile, updatePackageJsonFileAsync.
6. Updated build-cli lint config to only warn on uses of deprecated APIs
_except_ for vnext code.


[AB#19713](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/19713)