refactor(telemetry-generator): Use pnpm (#23327)
## Description

Telemetry-generator was the last bit of the repo still using `npm`
instead of `pnpm`. This fixes that and updates all places where we use
it to reflect the change.

Took the chance to update its dependencies to more recent versions,
syncing some (e.g. typescript) with the version we use in the client
release group.