test(client-presence): refactor of presence manager unit tests (#23300)
## Description

This PR refactors the unit tests in presenceManager.spec.ts in response
to [this
comment](https://github.com/microsoft/FluidFramework/pull/23113#discussion_r1861350670).
We now include a simulateAttendeeJoin helper function that can take in
multiple client join signals and returns the joined attendees while
maintaining order. This is to accomodate for future, more complicated
test scenarios where we simulate multiple joining clients, as well as
simulating join responses.

To show the benefit of the refactor I added a failing test case "is
announced via `attendeeJoined` when a second joining attendee is unknown
to audience" where two clients join and one responds to the other with
it's known information. Using the helper we simply send the signals in
order then verify the joinedAttendees array using another helper
verifyAttendee.