fix(offline): Mitigate flaky test by opting out of ID Compression for now (#23356)
The test `Parallel Forks: Closes (ForkedContainerError and
DuplicateBatchError) when hydrating twice and submitting in parallel
(via Counter DDS)` fails intermittently (but regularly) against t9s and
FRS. The test failure indicates that sometimes Container1 (an innocent
bystander observing the action) doesn't process the counter's increment
op.

Turns out, Container1 is closing with an unexpected error 100% of the
time - and there's some non-determinism, so sometimes it hits that error
before processing the counter's increment op.

The unexpected error is "Ranges finalized out of order" from the ID
Compressor. We can mitigate this by commenting out the line that
generates a compressed ID.

We're tracking the true bug fix elsewhere, this will remove the flakiness for now.