refactor(api-markdown-documenter): Unify configuration patterns (#23374)
This library has an inconsistent mix of `Partial` and `Required` types
to represent partial user input parameters and "complete" configurations
needed by the system to function.

This PR begins an attempt to align its APIs with the following
conventions:

-   Naming:
- "Options": refers to user-provided API parameters, which may be
incomplete.
- "Configuration": refers to the "complete" sets of parameters needed by
system functionality.
-   Typing:
- When possible, "configuration" types will be declared with all
properties required.
- When possible, "options" types will be declared as
`Partial<FooConfiguration>`. When not possible, they will be declared as
separate types.

Specifically, this PR replaces the `ApiItemTransformationConfiguration`
type with `ApiItemTransformationOptions` and
`ApiItemTransformationConfiguration`.

It also renames the existing `ApiItemTransformationOptions` type to
`ApiItemTransformations`, which better aligns it with similar concepts
in the renderers.