Precursor Work for Segment Deprecation Removals; Unify on Infos. (#23371)
This change aligns the insertion information of an ISegment with the
existing IRemovalInfo and IMoveInfo by adding a new interface,
IInsertionInfo. This work is primarily precursor work, which is related
to #23323, as these changes were made with some of those deprecation's
removed locally so that we can reduce the eventual churn when we finally
remove those properties.

In addition to adding the new interface, this change also standardizes
the methods we provide for each of these interfaces;
- toInsertionInfo/toRemovealInfo/toMoveInfo
- isInserted/isRemoved/IsMoved
- assertInserted/assertRemoved/assertMoved

There are also some general utility types and methods which make the
info types easier to work with:
- SegmentInfo
- SegmentWithInfo
- overwriteSegmentInfo
After deprecations are removed. These types will also aid in improving
typing to reduce the need for assertions.

Lastly, I moved all the info from ISegmentInternal to ISegmentLeaf. This
better approximates the final state, and allow further reducing the
exposure of the internal state of the segment. Due to this some changes
in other packages were necessary, but all could be solved with the
existing segmentIsRemovedFunction

---------

Co-authored-by: jzaffiro <110866475+jzaffiro@users.noreply.github.com>