refactor(api-markdown-documenter): Add and use a `ValidApiItemKind` type alias which excludes the `None` value (#23384)
No valid `ApiItem` will have kind "None". There are a number of places
where we are forced to assert that an item's kind is not "None" to make
checks exhaustive, and there are other places where we probably _should_
have been.

This PR adds a helper type alias and a helper function to encapsulate
the validity checking.