refactor(api-markdown-documenter): Rename `DocumentationSuiteOptions` to `DocumentationSuiteConfiguration` and make its properties required (#23388)
This library has an inconsistent mix of `Partial` and `Required` types
to represent partial user input parameters and "complete" configurations
needed by the system to function.
This version of the library attempts to align its APIs with the
following conventions:
-   Naming:
- "Options": refers to user-provided API parameters, which may be
incomplete.
- "Configuration": refers to the "complete" sets of parameters needed by
system functionality.
-   Typing:
- When possible, "configuration" types will be declared with all
properties required.
- When possible, "options" types will be declared as
`Partial<FooConfiguration>`. When not possible, they will be declared as
separate types.

This PR specifically makes the following name / structure changes:

- `DocumentationSuiteOptions` -> `DocumentationSuiteConfiguration` (user
input is taken as `Partial<DocumentationSuiteConfiguration>`).