test(client-presence): old collateral connection and stale connection tests (#23351)
## Description

There are 6 tests that are added (and disabled) in this PR. 

The first test is to verify that old collateral connections do not
effect the connection status of remote attendees. This scenario was
written as an edge case to [old
solution](https://github.com/microsoft/FluidFramework/pull/23113#discussion_r1844596112)
where old connection information sent in a join response would result in
the attendee connection status being set to 'Disconnected'.

The next five have to do with stale remote attendee's connection status
upon local client reconnect scenarios:

```
Presence
  PresenceManager
    when connected
      attendee
        that is joining
          - as collateral with old connection info and connected is NOT announced via `attendeeJoined
        that is already known
          and then local client disconnects
            - updates status of attendee with stale connection after 30s delay upon local reconnection
            - does not update status of attendee with stale connection if local client does not reconnect
            - does not update status of attendee with stale connection if local client reconnection lasts less than 30s
            - does not update status of attendee with stale connection if attendee rejoins
            - does not update status of attendee with stale connection if attendee sends datastore update
            - marks attendee with stale connection as active when attendee disconnects after local reconnection
            - updates status of attendee with stale connection only 30s after most recent local reconnection
```

I added connect() and disconnect() function to MockEphemeralRuntime that
manipulates the connection state of the local client.