MergeTree: Remove SegmentGroup from internal (#23401)
This change removes SegmentGroup, and a few other types from the
internal api. This exposure was unnecessary, as internal consumer just
need `localOpMetadata` which is an `unknown`, so the existing internal
consumers do not need the type. Additionally, this change will make some
additional refactoring around deprecated segment properties easier to
stage.