MergeTree: Remove Partial IRemovalInfo and IMoveInfo from Segments (#23406)
This is further refactoring which will aid in the eventual removal of
the deprecated properties from ISegment. This change removed the partial
implementation of IRemovalInfo and IMoveInfo from ISegmentPrivate which
forces consistent access to those infos via the typeguards and helpers
for the associated infos. This consistent access also ensures strong
typing, and correct handling for optional vs required properties.